Flask_UI/UI/FlaskApp/requirements.txt
kafka-python==2.0.2
PyYAML==6.0
Flask==2.0.2
jsonpickle==2.0.0
cassandra-driver==3.25.0